REF: fix import paths after move, adjust docs
